fix; post.ts

+++ src/routes/profiles/index.ts
@@ -38,6 +38,20 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
     },
     async function (request, reply): Promise<ProfileEntity> {
       return await fastify.db.profiles.create(request.body);
+      // const type = await fastify.db.memberTypes.findOne({
+      //   key: 'id',
+      //   equals: request.body.memberTypeId,
+      // });
+      // if (!type) throw fastify.httpErrors.createError(404, 'Type not found');
+
+      // const profile = await fastify.db.profiles.findOne({
+      //   key: 'userId',
+      //   equals: request.body.userId,
+      // });
+
+      // if (!profile) return await fastify.db.profiles.create(request.body);
+
+      // throw fastify.httpErrors.createError(404, 'Profile already exist');
     }
   );
 
@@ -49,6 +63,13 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
+      const profile = await fastify.db.profiles.findOne({
+        key: 'id',
+        equals: request.params.id,
+      });
+      if (!profile)
+        throw fastify.httpErrors.createError(400, 'Profile not found');
+
       return await fastify.db.profiles.delete(request.params.id);
     }
   );
